move extractors and generic schema utilities out of handler.rs
