+++ dropshot/src/handler.rs
@@ -32,34 +32,29 @@
 //! OpenAPI document generation.
 
 use super::error::HttpError;
-use super::http_util::http_extract_path_params;
-use super::http_util::http_read_body;
+use super::extractor::Extractor;
 use super::http_util::CONTENT_TYPE_JSON;
 use super::http_util::CONTENT_TYPE_OCTET_STREAM;
 use super::server::DropshotState;
 use super::server::ServerContext;
 use crate::api_description::ApiEndpointHeader;
-use crate::api_description::ApiEndpointParameter;
-use crate::api_description::ApiEndpointParameterLocation;
 use crate::api_description::ApiEndpointResponse;
 use crate::api_description::ApiSchemaGenerator;
-use crate::api_description::{ApiEndpointBodyContentType, ExtensionMode};
+use crate::api_description::{ApiEndpointBodyContentType};
 use crate::pagination::PaginationParams;
-use crate::pagination::PAGINATION_PARAM_SENTINEL;
 use crate::router::VariableSet;
+use crate::schema_util::make_subschema_for;
+use crate::schema_util::schema2struct;
+use crate::schema_util::ReferenceVisitor;
 use crate::to_map::to_map;
-use crate::websocket::WEBSOCKET_PARAM_SENTINEL;
 
 use async_trait::async_trait;
-use bytes::Bytes;
 use futures::lock::Mutex;
 use http::HeaderMap;
 use http::StatusCode;
 use hyper::Body;
 use hyper::Request;
 use hyper::Response;
-use schemars::schema::InstanceType;
-use schemars::schema::SchemaObject;
 use schemars::JsonSchema;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
@@ -150,77 +145,6 @@ impl<T: 'static + ServerContext> RequestContextArgument
 {
     type Context = T;
 }
-
-/// `Extractor` defines an interface allowing a type to be constructed from a
-/// `RequestContext`.  Unlike most traits, `Extractor` essentially defines only a
-/// constructor function, not instance functions.
-///
-/// The extractors that we provide (`Query`, `Path`, `TypedBody`, `UntypedBody`, and
-/// `WebsocketUpgrade`) implement `Extractor` in order to construct themselves from
-/// the request. For example, `Extractor` is implemented for `Query<Q>` with a
-/// function that reads the query string from the request, parses it, and
-/// constructs a `Query<Q>` with it.
-///
-/// We also define implementations of `Extractor` for tuples of types that
-/// themselves implement `Extractor`.  See the implementation of
-/// `HttpRouteHandler` for more on why this needed.
-#[async_trait]
-pub trait Extractor: Send + Sync + Sized {
-    /// Construct an instance of this type from a `RequestContext`.
-    async fn from_request<Context: ServerContext>(
-        rqctx: Arc<RequestContext<Context>>,
-    ) -> Result<Self, HttpError>;
-
-    fn metadata(
-        body_content_type: ApiEndpointBodyContentType,
-    ) -> ExtractorMetadata;
-}
-
-/// Metadata associated with an extractor including parameters and whether or not
-/// the associated endpoint is paginated.
-pub struct ExtractorMetadata {
-    pub extension_mode: ExtensionMode,
-    pub parameters: Vec<ApiEndpointParameter>,
-}
-
-/// `impl_derived_for_tuple!` defines implementations of `Extractor` for tuples
-/// whose elements themselves implement `Extractor`.
-macro_rules! impl_extractor_for_tuple {
-    ($( $T:ident),*) => {
-    #[async_trait]
-    impl< $($T: Extractor + 'static,)* > Extractor for ($($T,)*)
-    {
-        async fn from_request<Context: ServerContext>(_rqctx: Arc<RequestContext<Context>>)
-            -> Result<( $($T,)* ), HttpError>
-        {
-            futures::try_join!($($T::from_request(Arc::clone(&_rqctx)),)*)
-        }
-
-        fn metadata(_body_content_type: ApiEndpointBodyContentType) -> ExtractorMetadata {
-            #[allow(unused_mut)]
-            let mut extension_mode = ExtensionMode::None;
-            #[allow(unused_mut)]
-            let mut parameters = vec![];
-            $(
-                let mut metadata = $T::metadata(_body_content_type.clone());
-                extension_mode = match (extension_mode, metadata.extension_mode) {
-                    (ExtensionMode::None, x) | (x, ExtensionMode::None) => x,
-                    (x, y) if x != y => {
-                        panic!("incompatible extension modes in tuple: {:?} != {:?}", x, y);
-                    }
-                    (_, x) => x,
-                };
-                parameters.append(&mut metadata.parameters);
-            )*
-            ExtractorMetadata { extension_mode, parameters }
-        }
-    }
-}}
-
-impl_extractor_for_tuple!();
-impl_extractor_for_tuple!(T1);
-impl_extractor_for_tuple!(T1, T2);
-impl_extractor_for_tuple!(T1, T2, T3);
 
 /// `HttpHandlerFunc` is a trait providing a single function, `handle_request()`,
 /// which takes an HTTP request and produces an HTTP response (or
@@ -497,542 +421,6 @@ where
     }
 }
 
-// Extractors
-
-// Query: query string extractor
-
-/// `Query<QueryType>` is an extractor used to deserialize an instance of
-/// `QueryType` from an HTTP request's query string.  `QueryType` is any
-/// structure of yours that implements `serde::Deserialize`.  See this module's
-/// documentation for more information.
-#[derive(Debug)]
-pub struct Query<QueryType: DeserializeOwned + JsonSchema + Send + Sync> {
-    inner: QueryType,
-}
-
-impl<QueryType: DeserializeOwned + JsonSchema + Send + Sync> Query<QueryType> {
-    // TODO drop this in favor of Deref?  + Display and Debug for convenience?
-    pub fn into_inner(self) -> QueryType {
-        self.inner
-    }
-}
-
-/// Given an HTTP request, pull out the query string and attempt to deserialize
-/// it as an instance of `QueryType`.
-fn http_request_load_query<QueryType>(
-    request: &Request<Body>,
-) -> Result<Query<QueryType>, HttpError>
-where
-    QueryType: DeserializeOwned + JsonSchema + Send + Sync,
-{
-    let raw_query_string = request.uri().query().unwrap_or("");
-    // TODO-correctness: are query strings defined to be urlencoded in this way?
-    match serde_urlencoded::from_str(raw_query_string) {
-        Ok(q) => Ok(Query { inner: q }),
-        Err(e) => Err(HttpError::for_bad_request(
-            None,
-            format!("unable to parse query string: {}", e),
-        )),
-    }
-}
-
-// The `Extractor` implementation for Query<QueryType> describes how to construct
-// an instance of `Query<QueryType>` from an HTTP request: namely, by parsing
-// the query string to an instance of `QueryType`.
-// TODO-cleanup We shouldn't have to use the "'static" bound on `QueryType`
-// here.  It seems like we ought to be able to use 'async_trait, but that
-// doesn't seem to be defined.
-#[async_trait]
-impl<QueryType> Extractor for Query<QueryType>
-where
-    QueryType: JsonSchema + DeserializeOwned + Send + Sync + 'static,
-{
-    async fn from_request<Context: ServerContext>(
-        rqctx: Arc<RequestContext<Context>>,
-    ) -> Result<Query<QueryType>, HttpError> {
-        let request = rqctx.request.lock().await;
-        http_request_load_query(&request)
-    }
-
-    fn metadata(
-        _body_content_type: ApiEndpointBodyContentType,
-    ) -> ExtractorMetadata {
-        get_metadata::<QueryType>(&ApiEndpointParameterLocation::Query)
-    }
-}
-
-// Path: path parameter string extractor
-
-/// `Path<PathType>` is an extractor used to deserialize an instance of
-/// `PathType` from an HTTP request's path parameters.  `PathType` is any
-/// structure of yours that implements `serde::Deserialize`.  See this module's
-/// documentation for more information.
-#[derive(Debug)]
-pub struct Path<PathType: JsonSchema + Send + Sync> {
-    inner: PathType,
-}
-
-impl<PathType: JsonSchema + Send + Sync> Path<PathType> {
-    // TODO drop this in favor of Deref?  + Display and Debug for convenience?
-    pub fn into_inner(self) -> PathType {
-        self.inner
-    }
-}
-
-// The `Extractor` implementation for Path<PathType> describes how to construct
-// an instance of `Path<QueryType>` from an HTTP request: namely, by extracting
-// parameters from the query string.
-#[async_trait]
-impl<PathType> Extractor for Path<PathType>
-where
-    PathType: DeserializeOwned + JsonSchema + Send + Sync + 'static,
-{
-    async fn from_request<Context: ServerContext>(
-        rqctx: Arc<RequestContext<Context>>,
-    ) -> Result<Path<PathType>, HttpError> {
-        let params: PathType = http_extract_path_params(&rqctx.path_variables)?;
-        Ok(Path { inner: params })
-    }
-
-    fn metadata(
-        _body_content_type: ApiEndpointBodyContentType,
-    ) -> ExtractorMetadata {
-        get_metadata::<PathType>(&ApiEndpointParameterLocation::Path)
-    }
-}
-
-/// Convenience function to generate parameter metadata from types implementing
-/// `JsonSchema` for use with `Query` and `Path` `Extractors`.
-fn get_metadata<ParamType>(
-    loc: &ApiEndpointParameterLocation,
-) -> ExtractorMetadata
-where
-    ParamType: JsonSchema,
-{
-    // Generate the type for `ParamType` then pluck out each member of
-    // the structure to encode as an individual parameter.
-    let mut generator = schemars::gen::SchemaGenerator::new(
-        schemars::gen::SchemaSettings::openapi3(),
-    );
-    let schema = generator.root_schema_for::<ParamType>().schema.into();
-
-    let extension_mode = match schema_extensions(&schema) {
-        Some(extensions) => {
-            let paginated = extensions
-                .get(&PAGINATION_PARAM_SENTINEL.to_string())
-                .is_some();
-            let websocket =
-                extensions.get(&WEBSOCKET_PARAM_SENTINEL.to_string()).is_some();
-            match (paginated, websocket) {
-                (false, false) => ExtensionMode::None,
-                (false, true) => ExtensionMode::Websocket,
-                (true, false) => ExtensionMode::Paginated,
-                (true, true) => panic!(
-                    "Cannot use websocket and pagination in the same endpoint!"
-                ),
-            }
-        }
-        None => ExtensionMode::None,
-    };
-
-    // Convert our collection of struct members list of parameters.
-    let parameters = schema2struct(&schema, &generator, true)
-        .into_iter()
-        .map(|struct_member| {
-            let mut s = struct_member.schema;
-            let mut visitor = ReferenceVisitor::new(&generator);
-            schemars::visit::visit_schema(&mut visitor, &mut s);
-
-            ApiEndpointParameter::new_named(
-                loc,
-                struct_member.name,
-                struct_member.description,
-                struct_member.required,
-                ApiSchemaGenerator::Static {
-                    schema: Box::new(s),
-                    dependencies: visitor.dependencies(),
-                },
-                Vec::new(),
-            )
-        })
-        .collect::<Vec<_>>();
-
-    ExtractorMetadata { extension_mode, parameters }
-}
-
-fn schema_extensions(
-    schema: &schemars::schema::Schema,
-) -> Option<&schemars::Map<String, serde_json::Value>> {
-    match schema {
-        schemars::schema::Schema::Bool(_) => None,
-        schemars::schema::Schema::Object(object) => Some(&object.extensions),
-    }
-}
-
-/// Used to visit all schemas and collect all dependencies.
-struct ReferenceVisitor<'a> {
-    generator: &'a schemars::gen::SchemaGenerator,
-    dependencies: indexmap::IndexMap<String, schemars::schema::Schema>,
-}
-
-impl<'a> ReferenceVisitor<'a> {
-    fn new(generator: &'a schemars::gen::SchemaGenerator) -> Self {
-        Self { generator, dependencies: indexmap::IndexMap::new() }
-    }
-
-    fn dependencies(
-        self,
-    ) -> indexmap::IndexMap<String, schemars::schema::Schema> {
-        self.dependencies
-    }
-}
-
-impl<'a> schemars::visit::Visitor for ReferenceVisitor<'a> {
-    fn visit_schema_object(&mut self, schema: &mut SchemaObject) {
-        if let Some(refstr) = &schema.reference {
-            let definitions_path = &self.generator.settings().definitions_path;
-            let name = &refstr[definitions_path.len()..];
-
-            if !self.dependencies.contains_key(name) {
-                let mut refschema = self
-                    .generator
-                    .definitions()
-                    .get(name)
-                    .expect("invalid reference")
-                    .clone();
-                self.dependencies.insert(
-                    name.to_string(),
-                    schemars::schema::Schema::Bool(false),
-                );
-                schemars::visit::visit_schema(self, &mut refschema);
-                self.dependencies.insert(name.to_string(), refschema);
-            }
-        }
-
-        schemars::visit::visit_schema_object(self, schema);
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct StructMember {
-    pub name: String,
-    pub description: Option<String>,
-    pub schema: schemars::schema::Schema,
-    pub required: bool,
-}
-
-/// This helper function produces a list of the structure members for the
-/// given schema. For each it returns:
-///   (name: &String, schema: &Schema, required: bool)
-///
-/// If the input schema is not a flat structure the result will be a runtime
-/// failure reflective of a programming error (likely an invalid type specified
-/// in a handler function).
-///
-/// This function is invoked recursively on subschemas.
-pub(crate) fn schema2struct(
-    schema: &schemars::schema::Schema,
-    generator: &schemars::gen::SchemaGenerator,
-    required: bool,
-) -> Vec<StructMember> {
-    // We ignore schema.metadata, which includes things like doc comments, and
-    // schema.extensions. We call these out explicitly rather than eliding them
-    // as .. since we match all other fields in the structure.
-    match schema {
-        // We expect references to be on their own.
-        schemars::schema::Schema::Object(schemars::schema::SchemaObject {
-            metadata: _,
-            instance_type: None,
-            format: None,
-            enum_values: None,
-            const_value: None,
-            subschemas: None,
-            number: None,
-            string: None,
-            array: None,
-            object: None,
-            reference: Some(_),
-            extensions: _,
-        }) => schema2struct(
-            generator.dereference(schema).expect("invalid reference"),
-            generator,
-            required,
-        ),
-
-        // Match objects and subschemas.
-        schemars::schema::Schema::Object(schemars::schema::SchemaObject {
-            metadata: _,
-            instance_type: Some(schemars::schema::SingleOrVec::Single(_)),
-            format: None,
-            enum_values: None,
-            const_value: None,
-            subschemas,
-            number: None,
-            string: None,
-            array: None,
-            object,
-            reference: None,
-            extensions: _,
-        }) => {
-            let mut results = Vec::new();
-
-            // If there's a top-level object, add its members to the list of
-            // parameters.
-            if let Some(object) = object {
-                results.extend(object.properties.iter().map(
-                    |(name, schema)| {
-                        let (description, schema) =
-                            schema_extract_description(schema);
-                        StructMember {
-                            name: name.clone(),
-                            description,
-                            schema,
-                            required: required
-                                && object.required.contains(name),
-                        }
-                    },
-                ));
-            }
-
-            // We might see subschemas here in the case of flattened enums
-            // or flattened structures that have associated doc comments.
-            if let Some(subschemas) = subschemas {
-                match subschemas.as_ref() {
-                    // We expect any_of in the case of an enum.
-                    schemars::schema::SubschemaValidation {
-                        all_of: None,
-                        any_of: Some(schemas),
-                        one_of: None,
-                        not: None,
-                        if_schema: None,
-                        then_schema: None,
-                        else_schema: None,
-                    } => results.extend(schemas.iter().flat_map(|subschema| {
-                        // Note that these will be tagged as optional.
-                        schema2struct(subschema, generator, false)
-                    })),
-
-                    // With an all_of, there should be a single element. We
-                    // typically see this in the case where there is a doc
-                    // comment on a structure as OpenAPI 3.0.x doesn't have
-                    // a description field directly on schemas.
-                    schemars::schema::SubschemaValidation {
-                        all_of: Some(subschemas),
-                        any_of: None,
-                        one_of: None,
-                        not: None,
-                        if_schema: None,
-                        then_schema: None,
-                        else_schema: None,
-                    } if subschemas.len() == 1 => results.extend(
-                        subschemas.iter().flat_map(|subschema| {
-                            schema2struct(subschema, generator, required)
-                        }),
-                    ),
-
-                    // We don't expect any other types of subschemas.
-                    invalid => panic!("invalid subschema {:#?}", invalid),
-                }
-            }
-
-            results
-        }
-
-        // The generated schema should be an object.
-        invalid => panic!("invalid type {:#?}", invalid),
-    }
-}
-
-// TypedBody: body extractor for formats that can be deserialized to a specific
-// type.  Only JSON is currently supported.
-
-/// `TypedBody<BodyType>` is an extractor used to deserialize an instance of
-/// `BodyType` from an HTTP request body.  `BodyType` is any structure of yours
-/// that implements `serde::Deserialize`.  See this module's documentation for
-/// more information.
-#[derive(Debug)]
-pub struct TypedBody<BodyType: JsonSchema + DeserializeOwned + Send + Sync> {
-    inner: BodyType,
-}
-
-impl<BodyType: JsonSchema + DeserializeOwned + Send + Sync>
-    TypedBody<BodyType>
-{
-    // TODO drop this in favor of Deref?  + Display and Debug for convenience?
-    pub fn into_inner(self) -> BodyType {
-        self.inner
-    }
-}
-
-/// Given an HTTP request, attempt to read the body, parse it according
-/// to the content type, and deserialize it to an instance of `BodyType`.
-async fn http_request_load_body<Context: ServerContext, BodyType>(
-    rqctx: Arc<RequestContext<Context>>,
-) -> Result<TypedBody<BodyType>, HttpError>
-where
-    BodyType: JsonSchema + DeserializeOwned + Send + Sync,
-{
-    let server = &rqctx.server;
-    let mut request = rqctx.request.lock().await;
-    let body = http_read_body(
-        request.body_mut(),
-        server.config.request_body_max_bytes,
-    )
-    .await?;
-
-    // RFC 7231 §3.1.1.1: media types are case insensitive and may
-    // be followed by whitespace and/or a parameter (e.g., charset),
-    // which we currently ignore.
-    let content_type = request
-        .headers()
-        .get(http::header::CONTENT_TYPE)
-        .map(|hv| {
-            hv.to_str().map_err(|e| {
-                HttpError::for_bad_request(
-                    None,
-                    format!("invalid content type: {}", e),
-                )
-            })
-        })
-        .unwrap_or(Ok(CONTENT_TYPE_JSON))?;
-    let end = content_type.find(';').unwrap_or_else(|| content_type.len());
-    let mime_type = content_type[..end].trim_end().to_lowercase();
-    let body_content_type =
-        ApiEndpointBodyContentType::from_mime_type(&mime_type)
-            .map_err(|e| HttpError::for_bad_request(None, e))?;
-    let expected_content_type = rqctx.body_content_type.clone();
-
-    use ApiEndpointBodyContentType::*;
-    let content: BodyType = match (expected_content_type, body_content_type) {
-        (Json, Json) => serde_json::from_slice(&body).map_err(|e| {
-            HttpError::for_bad_request(
-                None,
-                format!("unable to parse JSON body: {}", e),
-            )
-        })?,
-        (UrlEncoded, UrlEncoded) => serde_urlencoded::from_bytes(&body)
-            .map_err(|e| {
-                HttpError::for_bad_request(
-                    None,
-                    format!("unable to parse URL-encoded body: {}", e),
-                )
-            })?,
-        (expected, requested) => {
-            return Err(HttpError::for_bad_request(
-                None,
-                format!(
-                    "expected content type \"{}\", got \"{}\"",
-                    expected.mime_type(),
-                    requested.mime_type()
-                ),
-            ))
-        }
-    };
-    Ok(TypedBody { inner: content })
-}
-
-// The `Extractor` implementation for TypedBody<BodyType> describes how to
-// construct an instance of `TypedBody<BodyType>` from an HTTP request: namely,
-// by reading the request body and parsing it as JSON into type `BodyType`.
-// TODO-cleanup We shouldn't have to use the "'static" bound on `BodyType` here.
-// It seems like we ought to be able to use 'async_trait, but that doesn't seem
-// to be defined.
-#[async_trait]
-impl<BodyType> Extractor for TypedBody<BodyType>
-where
-    BodyType: JsonSchema + DeserializeOwned + Send + Sync + 'static,
-{
-    async fn from_request<Context: ServerContext>(
-        rqctx: Arc<RequestContext<Context>>,
-    ) -> Result<TypedBody<BodyType>, HttpError> {
-        http_request_load_body(rqctx).await
-    }
-
-    fn metadata(content_type: ApiEndpointBodyContentType) -> ExtractorMetadata {
-        let body = ApiEndpointParameter::new_body(
-            content_type,
-            true,
-            ApiSchemaGenerator::Gen {
-                name: BodyType::schema_name,
-                schema: make_subschema_for::<BodyType>,
-            },
-            vec![],
-        );
-        ExtractorMetadata {
-            extension_mode: ExtensionMode::None,
-            parameters: vec![body],
-        }
-    }
-}
-
-// UntypedBody: body extractor for a plain array of bytes of a body.
-
-/// `UntypedBody` is an extractor for reading in the contents of the HTTP request
-/// body and making the raw bytes directly available to the consumer.
-#[derive(Debug)]
-pub struct UntypedBody {
-    content: Bytes,
-}
-
-impl UntypedBody {
-    /// Returns a byte slice of the underlying body content.
-    // TODO drop this in favor of Deref?  + Display and Debug for convenience?
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.content
-    }
-
-    /// Convenience wrapper to convert the body to a UTF-8 string slice,
-    /// returning a 400-level error if the body is not valid UTF-8.
-    pub fn as_str(&self) -> Result<&str, HttpError> {
-        std::str::from_utf8(self.as_bytes()).map_err(|e| {
-            HttpError::for_bad_request(
-                None,
-                format!("failed to parse body as UTF-8 string: {}", e),
-            )
-        })
-    }
-}
-
-#[async_trait]
-impl Extractor for UntypedBody {
-    async fn from_request<Context: ServerContext>(
-        rqctx: Arc<RequestContext<Context>>,
-    ) -> Result<UntypedBody, HttpError> {
-        let server = &rqctx.server;
-        let mut request = rqctx.request.lock().await;
-        let body_bytes = http_read_body(
-            request.body_mut(),
-            server.config.request_body_max_bytes,
-        )
-        .await?;
-        Ok(UntypedBody { content: body_bytes })
-    }
-
-    fn metadata(
-        _content_type: ApiEndpointBodyContentType,
-    ) -> ExtractorMetadata {
-        ExtractorMetadata {
-            parameters: vec![ApiEndpointParameter::new_body(
-                ApiEndpointBodyContentType::Bytes,
-                true,
-                ApiSchemaGenerator::Static {
-                    schema: Box::new(
-                        SchemaObject {
-                            instance_type: Some(InstanceType::String.into()),
-                            format: Some(String::from("binary")),
-                            ..Default::default()
-                        }
-                        .into(),
-                    ),
-                    dependencies: indexmap::IndexMap::default(),
-                },
-                vec![],
-            )],
-            extension_mode: ExtensionMode::None,
-        }
-    }
-}
-
 // Response Type Conversion
 //
 // See the discussion on macro `impl_HttpHandlerFunc_for_func_with_params` for a
@@ -1188,12 +576,6 @@ where
             ..Default::default()
         }
     }
-}
-
-fn make_subschema_for<T: JsonSchema>(
-    gen: &mut schemars::gen::SchemaGenerator,
-) -> schemars::schema::Schema {
-    gen.subschema_for::<T>()
 }
 
 /// `HttpResponseCreated<T: Serialize>` wraps an object of any serializable type.
@@ -1551,185 +933,5 @@ impl<
 
         metadata.headers = headers;
         metadata
-    }
-}
-
-fn schema_extract_description(
-    schema: &schemars::schema::Schema,
-) -> (Option<String>, schemars::schema::Schema) {
-    // Because the OpenAPI v3.0.x Schema cannot include a description with
-    // a reference, we may see a schema with a description and an `all_of`
-    // with a single subschema. In this case, we flatten the trivial subschema.
-    if let schemars::schema::Schema::Object(schemars::schema::SchemaObject {
-        metadata,
-        instance_type: None,
-        format: None,
-        enum_values: None,
-        const_value: None,
-        subschemas: Some(subschemas),
-        number: None,
-        string: None,
-        array: None,
-        object: None,
-        reference: None,
-        extensions: _,
-    }) = schema
-    {
-        if let schemars::schema::SubschemaValidation {
-            all_of: Some(subschemas),
-            any_of: None,
-            one_of: None,
-            not: None,
-            if_schema: None,
-            then_schema: None,
-            else_schema: None,
-        } = subschemas.as_ref()
-        {
-            match (subschemas.first(), subschemas.len()) {
-                (Some(subschema), 1) => {
-                    let description = metadata
-                        .as_ref()
-                        .and_then(|m| m.as_ref().description.clone());
-                    return (description, subschema.clone());
-                }
-                _ => (),
-            }
-        }
-    }
-
-    match schema {
-        schemars::schema::Schema::Bool(_) => (None, schema.clone()),
-
-        schemars::schema::Schema::Object(object) => {
-            let description = object
-                .metadata
-                .as_ref()
-                .and_then(|m| m.as_ref().description.clone());
-            (
-                description,
-                schemars::schema::SchemaObject {
-                    metadata: None,
-                    ..object.clone()
-                }
-                .into(),
-            )
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::api_description::ExtensionMode;
-    use crate::{
-        api_description::ApiEndpointParameterMetadata, ApiEndpointParameter,
-        ApiEndpointParameterLocation, PaginationParams,
-    };
-    use schemars::JsonSchema;
-    use serde::{Deserialize, Serialize};
-
-    use super::get_metadata;
-    use super::ExtractorMetadata;
-
-    #[derive(Deserialize, Serialize, JsonSchema)]
-    #[allow(dead_code)]
-    struct A {
-        foo: String,
-        bar: u32,
-        baz: Option<String>,
-    }
-
-    #[derive(JsonSchema)]
-    #[allow(dead_code)]
-    struct B<T> {
-        #[serde(flatten)]
-        page: T,
-
-        limit: Option<u64>,
-    }
-
-    #[derive(JsonSchema)]
-    #[allow(dead_code)]
-    #[schemars(untagged)]
-    enum C<T> {
-        First(T),
-        Next { page_token: String },
-    }
-
-    fn compare(
-        actual: ExtractorMetadata,
-        extension_mode: ExtensionMode,
-        parameters: Vec<(&str, bool)>,
-    ) {
-        assert_eq!(actual.extension_mode, extension_mode);
-
-        // This is order-dependent. We might not really care if the order
-        // changes, but it will be interesting to understand why if it does.
-        actual.parameters.iter().zip(parameters.iter()).for_each(
-            |(param, (name, required))| {
-                if let ApiEndpointParameter {
-                    metadata: ApiEndpointParameterMetadata::Path(aname),
-                    required: arequired,
-                    ..
-                } = param
-                {
-                    assert_eq!(aname, name);
-                    assert_eq!(arequired, required, "mismatched for {}", name);
-                } else {
-                    panic!();
-                }
-            },
-        );
-    }
-
-    #[test]
-    fn test_metadata_simple() {
-        let params = get_metadata::<A>(&ApiEndpointParameterLocation::Path);
-        let expected = vec![("bar", true), ("baz", false), ("foo", true)];
-
-        compare(params, ExtensionMode::None, expected);
-    }
-
-    #[test]
-    fn test_metadata_flattened() {
-        let params = get_metadata::<B<A>>(&ApiEndpointParameterLocation::Path);
-        let expected = vec![
-            ("bar", true),
-            ("baz", false),
-            ("foo", true),
-            ("limit", false),
-        ];
-
-        compare(params, ExtensionMode::None, expected);
-    }
-
-    #[test]
-    fn test_metadata_flattened_enum() {
-        let params =
-            get_metadata::<B<C<A>>>(&ApiEndpointParameterLocation::Path);
-        let expected = vec![
-            ("limit", false),
-            ("bar", false),
-            ("baz", false),
-            ("foo", false),
-            ("page_token", false),
-        ];
-
-        compare(params, ExtensionMode::None, expected);
-    }
-
-    #[test]
-    fn test_metadata_pagination() {
-        let params = get_metadata::<PaginationParams<A, A>>(
-            &ApiEndpointParameterLocation::Path,
-        );
-        let expected = vec![
-            ("bar", false),
-            ("baz", false),
-            ("foo", false),
-            ("limit", false),
-            ("page_token", false),
-        ];
-
-        compare(params, ExtensionMode::Paginated, expected);
     }
 }
