+++ dropshot/src/lib.rs
@@ -595,12 +595,14 @@ pub enum ProbeRegistration {
 mod api_description;
 mod config;
 mod error;
+mod extractor;
 mod from_map;
 mod handler;
 mod http_util;
 mod logging;
 mod pagination;
 mod router;
+mod schema_util;
 mod server;
 mod to_map;
 mod type_util;
@@ -627,11 +629,15 @@ pub use config::ConfigDropshot;
 pub use config::ConfigTls;
 pub use error::HttpError;
 pub use error::HttpErrorResponseBody;
+pub use extractor::Extractor;
+pub use extractor::ExtractorMetadata;
+pub use extractor::Path;
+pub use extractor::Query;
+pub use extractor::TypedBody;
+pub use extractor::UntypedBody;
 pub use handler::http_response_found;
 pub use handler::http_response_see_other;
 pub use handler::http_response_temporary_redirect;
-pub use handler::Extractor;
-pub use handler::ExtractorMetadata;
 pub use handler::FreeformBody;
 pub use handler::HttpCodedResponse;
 pub use handler::HttpResponse;
@@ -645,11 +651,7 @@ pub use handler::HttpResponseSeeOther;
 pub use handler::HttpResponseTemporaryRedirect;
 pub use handler::HttpResponseUpdatedNoContent;
 pub use handler::NoHeaders;
-pub use handler::Path;
-pub use handler::Query;
 pub use handler::RequestContext;
-pub use handler::TypedBody;
-pub use handler::UntypedBody;
 pub use http_util::CONTENT_TYPE_JSON;
 pub use http_util::CONTENT_TYPE_NDJSON;
 pub use http_util::CONTENT_TYPE_OCTET_STREAM;
